requirements.txt
absl-py==0.13.0
astunparse==1.6.3
atari-py==0.2.9
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
clang==5.0
cloudpickle==1.6.0
cycler==0.10.0
flatbuffers==1.12
gast==0.4.0
google-auth==1.34.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
grpcio==1.39.0
gym==0.18.3
gym-retro==0.8.0
h5py==3.1.0
idna==3.2
joblib==1.0.1
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.3
mpi4py==3.1.1
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.3.56
opt-einsum==3.3.0
pandas==1.3.2
Pillow==8.2.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyglet==1.5.15
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
retrowrapper==0.3.0
rsa==4.7.2
scipy==1.7.0
seaborn==0.11.2
six==1.15.0
stable-baselines3==1.1.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
torch==1.9.0
torchvision==0.10.0
typing-extensions==3.7.4.3
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
